Version: 2013-07-19 ksh93v-
Sourced from:
e40154a1dc8f3a8fd96dc181ea789b7f1e682381dd3725ada8701bddedd0b7df  ast-open.2013-07-19.tgz
6c2cca70db1926f6a209454ee74caf93fb7ba9a18388f35a64fe6a16d45a8e88  INIT.2013-07-19.tgz

+++ src/cmd/ksh93/include/jobs.h
@@ -96,7 +96,7 @@ struct process
 	unsigned short	p_exit;		/* exit value or signal number */
 	unsigned short	p_exitmin;	/* minimum exit value for xargs */
 	unsigned short	p_flag;		/* flags - see below */
-	int		p_env;		/* subshell environment number */
+	long		p_env;		/* subshell environment number */
 #ifdef JOBS
 	off_t		p_name;		/* history file offset for command */
 	struct termios	p_stty;		/* terminal state for job */
