+++ src/lib/libast/cdt/dtstrhash.c
@@ -37,7 +37,7 @@ uint dtstrhash(uint h, char* args, int n)
 #define	FNV_PRIME	((1<<24) + (1<<8) + 0x93)
 #define FNV_OFFSET	2166136261
 #endif
-	h = (h == 0 || h == ~0) ? FNV_OFFSET : h;
+	h = (h == 0 || h == (uint)~0) ? FNV_OFFSET : h;
 	if(n <= 0) /* see discipline key definition for == 0 */
 	{	for(; *s != 0; ++s )
 			h = ((h ^ s[0]) * FNV_PRIME) ^ (h >> (_ast_sizeof_int-1)*8);
